First version of clustering
Columns 41, 245, 246, 264 and 650 have been duplicated and checked for clusters. First I have split the columns on '^$' and when necessary also on '$ $ $' . Later I found out that it was also possible to first split the columns on '^$' and then split multivalued cells on '$ $ $'. After checking for clusters I rejoined the cells and/or columns. I also used a GREL function to remove whitespace at the beginning of the string (value.substring(1)) and to remove '^$' at the end of the string (value.chomp('^$' ).