Belgium-1601-1625_3.3
In this version, I tried to reconcile the publisher name in 264$ $1$b. First I dupliacate the column to prevent information loss. Then wanted to remove the string "by" in the front of the name using replace method and regular expression /^by\s/, but this doesn't apply to all cells starting with "by", so I did it manually with the cells that doesn't work. Finally, I started reconciling, but since there are a lot of multiple name in brackets and miss-spelling while transciption, the matching process doesn't have a effective outcome.