Add files via upload
In this version, I tried to split the column 246 to extract the author names, but that column contains the literal transcription of the (sub)title and author names so our amazing teammate suggested it may not be a good candidate for reconciling or splitting in this case. Then I undid it and turned to work on the column 700, 710 and their sub columns to use the Regular Expression to replace the historical name version like Iacobus and Ioannes/Iohannes into Jacobus and Johannes, the more modern version that may help the Wikidata to recognize them.