Belgium 1601-1625 4
Columns 041 and 246 have been duplicated and split. Column 146$3$3$a has been clustered. Column 245$0$0$a 3 was blank and has been removed. The data in column 245$0$4$a 3 belonged in another column, so it has been moved and then the empty column has been deleted. Column 700$0$ $a 1 1 1 has been renamed to column 700$0$ $a. Column 700$0$ $4 and column 700$1$ $4 have been renamed to $e and split in $3. ($4 is a UNIMARC or CNMARC tag, not a MARC21 tag). Column 710 had to be re-duplicated, -split, -clustered and -reconciled. According to the data model only the first indicator 2 should exist, so the other previous columns were faults in the data. 710$2$ $4 has been renamed to $e for the same previous reason.